MQE-2158: move proxies function from MagentoWebDriver back into MagentoActionProxies
