+++ src/Magento/FunctionalTestingFramework/Module/MagentoWebDriver.php
@@ -26,7 +26,6 @@ use Yandex\Allure\Adapter\AllureException;
 use Magento\FunctionalTestingFramework\DataTransport\Protocol\CurlTransport;
 use Yandex\Allure\Adapter\Support\AttachmentSupport;
 use Magento\FunctionalTestingFramework\Exceptions\TestFrameworkException;
-use Magento\FunctionalTestingFramework\DataGenerator\Handlers\PersistedObjectHandler;
 
 /**
  * MagentoWebDriver module provides common Magento web actions through Selenium WebDriver.
@@ -951,120 +950,6 @@ class MagentoWebDriver extends WebDriver
     public function getOTP()
     {
         return OTP::getOTP();
-    }
-
-    /**
-     * Create an entity
-     * TODO: move this function to MagentoActionProxies after MQE-1904
-     *
-     * @param string $key                 StepKey of the createData action.
-     * @param string $scope
-     * @param string $entity              Name of xml entity to create.
-     * @param array  $dependentObjectKeys StepKeys of other createData actions that are required.
-     * @param array  $overrideFields      Array of FieldName => Value of override fields.
-     * @param string $storeCode
-     * @return void
-     */
-    public function createEntity(
-        $key,
-        $scope,
-        $entity,
-        $dependentObjectKeys = [],
-        $overrideFields = [],
-        $storeCode = ''
-    ) {
-        PersistedObjectHandler::getInstance()->createEntity(
-            $key,
-            $scope,
-            $entity,
-            $dependentObjectKeys,
-            $overrideFields,
-            $storeCode
-        );
-    }
-
-    /**
-     * Retrieves and updates a previously created entity
-     * TODO: move this function to MagentoActionProxies after MQE-1904
-     *
-     * @param string $key                 StepKey of the createData action.
-     * @param string $scope
-     * @param string $updateEntity        Name of the static XML data to update the entity with.
-     * @param array  $dependentObjectKeys StepKeys of other createData actions that are required.
-     * @return void
-     */
-    public function updateEntity($key, $scope, $updateEntity, $dependentObjectKeys = [])
-    {
-        PersistedObjectHandler::getInstance()->updateEntity(
-            $key,
-            $scope,
-            $updateEntity,
-            $dependentObjectKeys
-        );
-    }
-
-    /**
-     * Performs GET on given entity and stores entity for use
-     * TODO: move this function to MagentoActionProxies after MQE-1904
-     *
-     * @param string  $key                 StepKey of getData action.
-     * @param string  $scope
-     * @param string  $entity              Name of XML static data to use.
-     * @param array   $dependentObjectKeys StepKeys of other createData actions that are required.
-     * @param string  $storeCode
-     * @param integer $index
-     * @return void
-     */
-    public function getEntity($key, $scope, $entity, $dependentObjectKeys = [], $storeCode = '', $index = null)
-    {
-        PersistedObjectHandler::getInstance()->getEntity(
-            $key,
-            $scope,
-            $entity,
-            $dependentObjectKeys,
-            $storeCode,
-            $index
-        );
-    }
-
-    /**
-     * Retrieves and deletes a previously created entity
-     * TODO: move this function to MagentoActionProxies after MQE-1904
-     *
-     * @param string $key   StepKey of the createData action.
-     * @param string $scope
-     * @return void
-     */
-    public function deleteEntity($key, $scope)
-    {
-        PersistedObjectHandler::getInstance()->deleteEntity($key, $scope);
-    }
-
-    /**
-     * Retrieves a field from an entity, according to key and scope given
-     * TODO: move this function to MagentoActionProxies after MQE-1904
-     *
-     * @param string $stepKey
-     * @param string $field
-     * @param string $scope
-     * @return string
-     */
-    public function retrieveEntityField($stepKey, $field, $scope)
-    {
-        return PersistedObjectHandler::getInstance()->retrieveEntityField($stepKey, $field, $scope);
-    }
-
-    /**
-     * Get encrypted value by key
-     * TODO: move this function to MagentoActionProxies after MQE-1904
-     *
-     * @param string $key
-     * @return string|null
-     * @throws TestFrameworkException
-     */
-    public function getSecret($key)
-    {
-        return CredentialStore::getInstance()->getSecret($key);
     }
 
     /**
